requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==1.10.13
starlette==0.27.0
typing-extensions==4.8.0
unidecode==1.3.7
